CMake Includes refactoring
