+++ CMakeLists.txt
@@ -11,3 +11,4 @@ add_subdirectory(dev)
 enable_testing()
 
 add_subdirectory(unit_test)
+
