+++ dev/Controller/CMakeLists.txt
@@ -10,3 +10,10 @@ set(CONTROLLER_HEADER_FILES
                     system/timer.h)
 
 add_library(Controller controller_main.cpp ${CONTROLLER_HEADER_FILES})
+
+target_include_directories(Controller 
+                           PUBLIC 
+                                ${CMAKE_CURRENT_SOURCE_DIR}/commands)
+                            
+
+target_link_libraries(Controller PUBLIC System_Interface Utilities)
