+++ dev/Controller/controller_main.cpp
@@ -1,7 +1,7 @@
 #include <algorithm>    // std::sort
 
 #include "controller_main.h"
-#include "../Interfaces/controller_interface.h"
+#include "Interfaces/controller_interface.h"
 
 
 Timer controller_timer;
