+++ dev/Controller/system/timer.h
@@ -1,6 +1,6 @@
-#ifndef MODEL_SYSTEM_VARIABLES
-#define MODEL_SYSTEM_VARIABLES
-#include "../../Utilities/timer.h"
+#ifndef CONTROLLER_SYSTEM_VARIABLES
+#define CONTROLLER_SYSTEM_VARIABLES
+#include "Utilities/timer.h"
 extern Timer controller_timer;
 
 #endif // !MODEL_SYSTEM_VARIABLES
