+++ dev/Model/CMakeLists.txt
@@ -21,5 +21,5 @@ set(MODEL_HEADER_FILES
 
 
 add_library(Model model_main.cpp ${MODEL_DEVICES_HEADER_FILES} ${MODEL_SYSTEM_HEADER_FILES} ${MODEL_STATE_HEADER_FILES} ${MODEL_NODE_HEADER_FILES} ${MODEL_HEADER_FILES})
-target_link_libraries(Model Utilities Interface)
+target_link_libraries(Model PUBLIC Utilities System_Interface)
 set_property(TARGET Model PROPERTY CXX_STANDARD 11)
