+++ dev/Model/model_main.h
@@ -11,7 +11,7 @@
 
 #include "system/timer.h"
 
-#include "../Interfaces/types/model_command.h"
+#include "Interfaces/types/model_command.h"
 
 /**
  * Main brain of the model. It keeps track of nodes and does actions such as looping, stepping through commands given from the Command_Interface, and sending nodes commands to run with devices.
