+++ dev/Model/system/timer.h
@@ -1,6 +1,6 @@
 #ifndef MODEL_SYSTEM_VARIABLES
 #define MODEL_SYSTEM_VARIABLES
-#include "../../Utilities/timer.h"
+#include "Utilities/timer.h"
 extern Timer model_timer;
 
 #endif // !MODEL_SYSTEM_VARIABLES
