+++ dev/Utilities/CMakeLists.txt
@@ -2,10 +2,22 @@
 set(UTILITIES_HEADER_FILES 
                     "Logging/log_item.h"
                     "Logging/logging.h"
-                    "save_reader.h"
-                    "timer.h"
-                    "utils.h"
-                    "system.h"
-                    "exceptions.h")
+                    "Utilities/save_reader.h"
+                    "Utilities/timer.h"
+                    "Utilities/utils.h"
+                    "Utilities/system.h"
+                    "Utilities/exceptions.h")
 
 add_library(Utilities Logging/logging.cpp ${UTILITIES_HEADER_FILES})
+target_include_directories(Utilities 
+                            PUBLIC
+                                "$<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}>"  
+                                "$<INSTALL_INTERFACE:${CMAKE_BINARY_DIR}/inc/System_Interface>")
+# Set the output directory for build artifacts
+  set_target_properties(Utilities
+    PROPERTIES
+    LIBRARY_OUTPUT_DIRECTORY "${CMAKE_BINARY_DIR}/lib/Utilities"
+    ARCHIVE_OUTPUT_DIRECTORY "${CMAKE_BINARY_DIR}/lib/Utilities"
+    PDB_OUTPUT_DIRECTORY "${CMAKE_BINARY_DIR}/bin")
+
+install(FILES ${INTERFACE_HEADER_FILES} DESTINATION ${CMAKE_BINARY_DIR}/inc/Utilities)
