+++ dev/View/Desktop/Interfaces/common/painter.h
@@ -1,6 +1,6 @@
 #ifndef COMMON_PAINTER_INTERFACE
 #define COMMON_PAINTER_INTERFACE
-#include "../../../../Interfaces/common_objects.h"
+#include "Interfaces/common_objects.h"
 #include "lib/dimensions.h"
 class Painter
 {
